+++ heim-process/tests/smoke.rs
@@ -1,5 +1,6 @@
 use heim_common::prelude::*;
 use heim_process as process;
+use heim_process::ProcessError;
 
 #[heim_derive::test]
 async fn smoke_pid_exists() {
@@ -17,6 +18,28 @@ async fn smoke_pids() {
     }
 }
 
+/// Try to .await the `Process` method and panic if there is a loading error occured.
+///
+/// Both successful results and
+/// NoSuchProcess/AccessDenied/ZombieProcess errors
+/// are going to be ignored, as we can't guarantee
+/// that queried process will be okay during the testing.
+macro_rules! try_method {
+    ($method:expr) => {
+        match $method.await {
+            Err(ProcessError::Load(e)) => {
+                assert!(
+                    false,
+                    "`{}` method returned an error: {:#?}",
+                    stringify!($method),
+                    e
+                );
+            }
+            _ => {}
+        }
+    };
+}
+
 #[heim_derive::test]
 async fn smoke_processes() {
     let mut processes = process::processes();
@@ -25,22 +48,20 @@ async fn smoke_processes() {
         let process = process.unwrap();
 
         let _ = process.pid();
-        let _ = process.parent_pid().await;
-        let _ = process.name().await;
-        let _ = process.exe().await;
-        let _ = process.status().await;
-        let _ = process.cpu_time().await;
-        let _ = process.memory().await;
+        try_method!(process.parent_pid());
+        try_method!(process.name());
+        try_method!(process.exe());
+        try_method!(process.status());
+        try_method!(process.cpu_time());
+        try_method!(process.memory());
+        try_method!(process.is_running());
 
         #[cfg(target_os = "linux")]
         {
             use heim_process::os::linux::ProcessExt;
 
-            let _ = process.io_counters().await;
-            let _ = process
-                .net_io_counters()
-                .try_for_each(|_| future::ok(()))
-                .await;
+            try_method!(process.io_counters());
+            try_method!(process.net_io_counters().try_for_each(|_| future::ok(())));
         }
     }
 }